Sources/server/requirements.txt
ray>=2.8.0
websockets>=11.0.3
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvloop>=0.19.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0 